Center Point Raster Template

Creates a SpatRaster template from a SpatVector point object in which the raster cells are centered on the vector points.

centerPointRasterTemplate(rawPointData)

Arguments

rawPointData

A SpatVector object with points that will represent the center of each cell in the output template.

Value

An empty SpatRaster designed to serve as a template for rasterizing SpatVector objects.

Details

rawPointData is a SpatVector object that contains x and y coordinates.

See also

Examples

library(terra)
#> terra 1.7.23
#> 
#> Attaching package: ‘terra’
#> The following objects are masked from ‘package:testthat’:
#> 
#>     compare, describe

# Create point grid
coords <- data.frame(x = rep(seq(1:5), times = 5),
                    y = unlist(lapply(1:5, FUN = function(x) {
                      rep(x, times = 5)})))

# Create data and add NAs to simulate uneven bottom depths
dd <- data.frame(SURFACE = 1:25,
                d5M = 6:30,
                d10M = 11:35,
                d25M = 16:40)
dd$d25M[c(1:5, 18:25)] <- NA
dd$d10M[c(3:5, 21:23)] <- NA
dd$d5M[c(4, 22)] <- NA

dd[,c("x","y")] <- coords

# Create SpatialPointsDataFrame
sp <- vect(dd, geom = c("x", "y"))

# Here's the function
template <- centerPointRasterTemplate(rawPointData = sp)
class(template)
#> [1] "SpatRaster"
#> attr(,"package")
#> [1] "terra"