Occurrence downsampling

Reduces number of occurrences to resolution of input raster

downsample(occs, rasterTemplate)

Arguments

occs
A data.frame with at least two columns named "longitude" and "latitude" or that can be coerced into this format.

rasterTemplate
A Raster* object to serve as a template for the resolution at which occs should be downsampled.

Value

A data.frame with two columns named "longitude" and "latitude" or with names that were used when coercing input data into this format.

Examples

# Create sample raster
r <- raster(ncol=10, nrow=10)

#> Error in raster(ncol = 10, nrow = 10): could not find function "raster"

values(r) <- 1:100

#> Error in values(r) <- 1:100: object 'r' not found

# Create test occurrences
set.seed(0)
longitude <- sample(extent(r)[1]:extent(r)[2],
                    size = 10, replace = FALSE)

#> Error in extent(r): could not find function "extent"

set.seed(0)
latitude <- sample(extent(r)[3]:extent(r)[4],
                   size = 10, replace = FALSE)

#> Error in extent(r): could not find function "extent"

# Here's the function
result <- downsample(occs = occurrences, rasterTemplate = r)

#> Error in downsample(occs = occurrences, rasterTemplate = r): could not find function "downsample"