Interpolate patchily sampled rasters

Uses thin plate spline regression from fields package to interpolate missing two-dimensional raster values.

interpolateRaster(inputRaster, fast = F, ...)

Arguments

inputRaster
An object of class raster

fast
A logical operator. Setting to T triggers use of fastTps instead of Tps.

...
For any additional arguments passed to Tps or fastTps

Value

An object of class raster

Details

Missing data values from original raster are replaced with interpolated values. User has the option of choosing fastTps to speed calculation, but be advised that this is only an approximation of a true thin plate spline.

See also

Examples

library(raster)
library(fields)

#> Loading required package: spam

#> Loading required package: dotCall64

#> Loading required package: grid

#> Spam version 2.7-0 (2021-06-25) is loaded.
#> Type 'help( Spam)' or 'demo( spam)' for a short introduction 
#> and overview of this package.
#> Help for individual functions is also obtained by adding the
#> suffix '.spam' to the function name, e.g. 'help( chol.spam)'.

#> 
#> Attaching package: ‘spam’

#> The following objects are masked from ‘package:base’:
#> 
#>     backsolve, forwardsolve

#> Loading required package: viridis

#> Loading required package: viridisLite

#> 
#> Try help(fields) to get started.

# Create sample raster
r <- raster(ncol=100, nrow=100)
values(r) <- 1:10000

# Introduce a "hole"
values(r)[67:70] <- NA

# Patch hole with interpolateRaster
interpolatedRaster <- interpolateRaster(r)

#> Warning: 
#> Grid searches over lambda (nugget and sill variances) with  minima at the endpoints: 
#>   (GCV) Generalized Cross-Validation 
#>    minimum at  right endpoint  lambda  =  3.025469e-06 (eff. df= 593.75 )

fastInterp <- interpolateRaster(r, fast = TRUE, aRange = 3.0)