Comparative raster mapping

A convenient wrapper around spplot to generate a formatted plot showing . This is used in the context of voluModel to overlay semi-transparent distributions (coded as 1) in two different RasterLayers.

oneRasterPlot(rast, land = NA, landCol = "black", title = "A Raster", ...)

Arguments

rast
A single Raster* layer on a continuous scale.

land
An optional coastline polygon shapefile of type sf to provide geographic context for the occurrence points.

landCol
Color for land on map.

title
A title for the plot.

...
Additional optional arguments to pass to spplot initial plot object or viridis.

Note

The extents of rast1 and rast2 must match.

See also

Examples

library(raster)
rast <- raster(ncol=10, nrow=10)
values(rast) <- seq(0,99, 1)

oneRasterPlot(rast = rast, )
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'obj' in selecting a method for function 'sppanel': no method or default for coercing “logical” to “Spatial”