Point mapping

A convenient wrapper around ggplot to generate formatted occurrence point plots.

pointMap(occs, spName, world)

Arguments

occs
A data.frame with at least two columns named "longitude" and "latitude" or that can be coerced into this format.

spName
A character string with the species name to be used in the plot title.

world
An optional coastline shapefile to provide geographic context for the occurrence points.

...
Additional optional arguments to pass to ggplot initial plot object.

Value

A data.frame with two columns named "longitude" and "latitude" or with names that were used when coercing input data into this format.

Details

The meat of this function is a special-case wrapper around getDynamicAlphaHull from the rangeBuilder package. The function documented here is especially useful in cases where one wants to automatically generate training regions that overlap the international date line. Regions that exceed the line are cut and pasted into the appropriate hemisphere instead of being deleted.

See also

Examples

library(rnaturalearth)

occs <- read.csv(system.file("extdata/", package='voluModel'))

#> Warning: 'raw = FALSE' but '/home/runner/work/_temp/Library/voluModel/extdata/' is not a regular file

#> Warning: cannot open file '/home/runner/work/_temp/Library/voluModel/extdata/': it is a directory

#> Error in file(file, "rt"): cannot open the connection

world <- ne_countries(scale = "medium", returnclass = "sf")

#> The rnaturalearthdata package needs to be installed.

#> Installing the rnaturalearthdata package.

#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)

#> Error in value[[3L]](cond): Failed to install the rnaturalearthdata package.
#>  Please try installing the package for yourself using the following command: 
#>  install.packages("rnaturalearthdata")