Sampling from rasterBrick using 3D coordinates

Gets values at x,y,z occurrences from a given 3D environmental variable brick

xyzSample(occs, envBrick)

Arguments

occs
A dataframe with at least three columns named "longitude", "latitude", and "depth", or that can be coerced into this format.

envBrick
A rasterBrick object with one environmental variable. Each layer represents a depth slice. See Details for more information.

Value

Vector of environmental values equal in length to number of rows of input occs data.frame.

Details

The envBrick rasterBrick object should have numeric names that correspond with the beginning depth of a particular depth slice. For example, one might have three layers, one from 0 to 10m, one from 10 to 30m, and one from 30 to 100m. You would name the layers in this brick names(envBrick) <- c(0, 10, 30. R will rename the layers "X0", "X10", and "X30". This is expected behavior and xyzSample was written to expect this. xyzSample identifies the layer name that is closest to the depth layer value at a particular X, Y coordinate, and samples the environmental value at that 3D coordinate.

Examples

library(raster)

# Create test raster
r1 <- raster(ncol=10, nrow=10)
values(r1) <- 1:100
r2 <- raster(ncol=10, nrow=10)
values(r2) <- c(rep(20, times = 50), rep(60, times = 50))
r3 <- raster(ncol=10, nrow=10)
values(r3) <- 8
envBrick <- brick(r1, r2, r3)
names(envBrick) <- c(0, 10, 30)

# Create test occurrences
set.seed(0)
longitude <- sample(extent(envBrick)[1]:extent(envBrick)[2],
                    size = 10, replace = FALSE)
set.seed(0)
latitude <- sample(extent(envBrick)[3]:extent(envBrick)[4],
                   size = 10, replace = FALSE)
set.seed(0)
depth <- sample(0:35, size = 10, replace = TRUE)
occurrences <- as.data.frame(cbind(longitude,latitude,depth))

# Test function
occSample3d <- xyzSample(occurrences, envBrick)