Occurrence downsampling

Reduces number of occurrences to resolution of input raster

downsample(occs, rasterTemplate)

Arguments

occs

A data.frame with at least two columns named "longitude" and "latitude" or that can be coerced into this format.

rasterTemplate

A Raster* object to serve as a template for the resolution at which occs should be downsampled.

Value

A data.frame with two columns named "longitude" and "latitude" or with names that were used when coercing input data into this format.

Examples

library(raster)
# Create sample raster
r <- raster(ncol=10, nrow=10)
values(r) <- 1:100

# Create test occurrences
set.seed(0)
longitude <- sample(extent(r)[1]:extent(r)[2],
                    size = 10, replace = FALSE)
set.seed(0)
latitude <- sample(extent(r)[3]:extent(r)[4],
                   size = 10, replace = FALSE)
occurrences <- as.data.frame(cbind(longitude,latitude))

# Here's the function
result <- downsample(occs = occurrences, rasterTemplate = r)
#> Using longitude and latitude
#>  as x and y coordinates, respectively.
head(result)
#>   longitude latitude
#> 1       126       45
#> 2       -18      -27
#> 3       -54       81
#> 4        90       63
#> 5        18      -45
#> 6       126      -81